+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 -   `Events.MOUSEWHEEL` and `Events.SCROLL` (which are the same), propagating mouse scroll events.
 -   `GameObject`s and `Group`s can now be hidden. In order to make their components or gos respectively not draw.
 -   `GameObjects`s can be active just like groups.
+-   `Component`s hidden properly implmented and does not stop update anymore.
 
 ### Changed